fix: do not open asset details in send flow
