+++ apps/mobile/src/features/balances/bitcoin/bitcoin-balance.tsx
@@ -1,12 +1,9 @@
 import { TokenBalance, TokenBalanceProps } from '@/features/token/components/token-balance';
 import { useBtcAccountBalance } from '@/queries/balance/btc-balance.query';
 import { t } from '@lingui/core/macro';
-import { useRouter } from 'expo-router';
 
-import { btcAsset } from '@leather.io/constants';
 import { AccountId } from '@leather.io/models';
 import { BitcoinFilledCircleIcon, BtcAvatarIcon } from '@leather.io/ui/native';
-import { getAssetId, serializeAssetId } from '@leather.io/utils';
 
 type BitcoinTokenBalanceProps = Omit<TokenBalanceProps, 'ticker' | 'tokenName' | 'icon'>;
 
@@ -21,12 +18,19 @@ export function BitcoinTokenBalance(props: BitcoinTokenBalanceProps) {
   );
 }
 
-export function BitcoinBalanceByAccount({ accountIndex, fingerprint }: AccountId) {
+interface BitcoinBalanceByAccountProps extends AccountId {
+  onPress?(): void;
+}
+
+export function BitcoinBalanceByAccount({
+  accountIndex,
+  fingerprint,
+  onPress,
+}: BitcoinBalanceByAccountProps) {
   const { state, value } = useBtcAccountBalance(fingerprint, accountIndex);
 
   const availableBalance = value?.btc.availableBalance;
   const quoteBalance = value?.quote.availableBalance;
-  const router = useRouter();
 
   if (!availableBalance || !quoteBalance) {
     return null;
@@ -35,12 +39,7 @@ export function BitcoinBalanceByAccount({ accountIndex, fingerprint }: AccountId
     <BitcoinTokenBalance
       availableBalance={availableBalance}
       quoteBalance={quoteBalance}
-      onPress={() =>
-        router.navigate({
-          pathname: '/(tabs)/(index)/[assetId]',
-          params: { assetId: serializeAssetId(getAssetId(btcAsset)) },
-        })
-      }
+      onPress={onPress}
       isLoading={state === 'loading'}
     />
   );
